Queue bug fix
Add Queue.hpp to Data Structure.  “The problem should be that you include the .cpp file in the main file. In C++, .cpp files are source files and should be compiled rather than included. When you include an .cpp file, the compiler copies the contents of that file directly to where it is included, which can cause you to define the same function multiple times in different files, throwing a link error” Doing so ensures that each function is defined only once in a .cpp file and does not raise linking errors.